dcbr-api/requirements.txt
# Main webapp
Django==2.2

# Database
psycopg2==2.8.2

# Web server
gunicorn>=19.9.0,<20
whitenoise>=4.1.2,<4.2

# REST framework
django-cors-headers==2.5.3
djangorestframework==3.9.3